remove the spec from the controller
this modifies the priority queue to just use the id.

we want the queue at some point, but it is just making it more difficult
in the interim. until we have some reason to have a queue (which I
believe should exist), there's really no reason to bother with it.